+++ stream-compositions/cursors/transaction-cursor/src/main/java/com/backbase/stream/compositions/transaction/cursor/core/domain/TransactionCursorEntity.java
@@ -1,11 +1,11 @@
 package com.backbase.stream.compositions.transaction.cursor.core.domain;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import java.sql.Timestamp;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
 
 import org.hibernate.annotations.GenericGenerator;