+++ stream-portfolio/portfolio-core/src/main/java/com/backbase/stream/portfolio/configuration/PortfolioSagaConfiguration.java
@@ -1,21 +1,21 @@
 package com.backbase.stream.portfolio.configuration;
 
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
+import com.backbase.stream.clients.autoconfigure.DbsApiClientsAutoConfiguration;
 import com.backbase.stream.portfolio.PortfolioSaga;
 import com.backbase.stream.portfolio.service.InstrumentIntegrationService;
 import com.backbase.stream.portfolio.service.PortfolioIntegrationService;
-import com.backbase.stream.webclient.configuration.DbsWebClientConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 
 /**
  * PortfolioSaga Configuration.
- * 
+ *
  * @author Vladimir Kirchev
  *
  */
 @Configuration
-@Import({DbsWebClientConfiguration.class})
+@Import(DbsApiClientsAutoConfiguration.class)
 public class PortfolioSagaConfiguration {
 
     @Bean